dt4sec-demo/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
paho-mqtt==2.1.0
aiosqlite==0.20.0
jinja2==3.1.4
scikit-learn==1.5.2
requests==2.32.3
numpy==2.1.1
